Test full user list
This provides some simple tests of the functionality that returns the full user list.

I also converted `User.age` to an `int`. We had this as an `Integer` which didn't
seem necessary, was a bit heavyweight, and complicated the use of `assertEquals()`
in the testing, so I changed it to an `int`.

Issue #8