Update `LABTASKS.md` (partially)
I've made it through most but not all of an update of `LABTASKS.md`, making
this line up better with the new lab.

There are (to me) a _ton_ of these questions for people to answer and turn
in, and it's not clear to me that (a) these are helpful to the students and (b) we
have the resources to read and respond to them in a timely fashion. I'd love
feedback from others on this.

I've gotten through documenting the API extensions, but I've written up
*nothing* about what has to be done on the client side to access those
features. There are also some unmodified questions at the end (after the
API description) that need to be read through and dealt with.

Issue #3