Add agile planning to write-up
This adds a somewhat lengthy write-up of using ZenHub to support agile planning, etc.

Given the length, one thing that isn't clear is whether we should move this from
LABTASKS to README. It's a description of process, so it's not "code you need to write",
but it is "stuff you need to do", so it's not obvious where it best goes. If people have
thoughts or suggestions on that, feel free to share.

Fixes #8